add cades sign format

+++ src/com/unical/digitalsignature/Main.java
@@ -1,6 +1,5 @@
 package com.unical.digitalsignature;
 
-import java.io.Console;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
@@ -10,25 +9,30 @@ import com.beust.jcommander.ParameterException;
 import com.google.common.io.Files;
 import com.google.common.io.Resources;
 import com.unical.utils.ArgsParser;
+import com.unical.utils.Utility;
 
+import eu.europa.esig.dss.AbstractSignatureParameters;
 import eu.europa.esig.dss.DSSDocument;
 import eu.europa.esig.dss.DSSException;
 import eu.europa.esig.dss.DigestAlgorithm;
 import eu.europa.esig.dss.FileDocument;
 import eu.europa.esig.dss.SignatureValue;
 import eu.europa.esig.dss.ToBeSigned;
-import eu.europa.esig.dss.pades.PAdESSignatureParameters;
-import eu.europa.esig.dss.pades.signature.PAdESService;
+import eu.europa.esig.dss.signature.AbstractSignatureService;
 import eu.europa.esig.dss.token.DSSPrivateKeyEntry;
 import eu.europa.esig.dss.token.Pkcs11SignatureToken;
 
 public class Main {
-	private static String separator = System.getProperty("file.separator");
-	private static File driverPathWin = new File(buildFilePath("driver", "Windows", "bit4xpki.dll"));
-	private static File driverPathLinux32 = new File(buildFilePath("driver", "Linux", "32", "libbit4xpki.so"));
-	private static File driverPathLinux64 = new File(buildFilePath("driver", "Linux", "64", "libbit4xpki.so"));
+	
+	private enum SystemType {WINDOWS,LINUX,MAC}
+
+	private static File driverPathWin = new File(Utility.buildFilePath("driver", "Windows", "bit4xpki.dll"));
+	private static File driverPathLinux32 = new File(Utility.buildFilePath("driver", "Linux", "32", "libbit4xpki.so"));
+	private static File driverPathLinux64 = new File(Utility.buildFilePath("driver", "Linux", "64", "libbit4xpki.so"));
 
 	private static File currentDriverPath = null;
+
+	private static SignFormat selectedSignFormat = SignFormat.PADES;
 
 	public static void main(String[] args) {
 		ArgsParser cmdr = new ArgsParser();
@@ -37,15 +41,22 @@ public class Main {
 		} catch (ParameterException e) {
 			System.err.println("Missing Parameter");
 			return;
-			// e.printStackTrace();
 		}
 		if (cmdr.isHelp()) {
 			cmdr.showHelp();
 			return;
 		}
-		if (cmdr.getDriver() != null)
-			setDriver(cmdr.getDriver());
-		else
+		if (!checkSelectedSignFormat(cmdr)) {
+			System.err.println("PAdES and CAdES are mutually exclusive. Select only one please.");
+			return;
+		}
+
+		if (cmdr.getDriver() != null) {
+			if (!setDriver(cmdr.getDriver())) {
+				System.err.println("Error setting driver");
+				return;
+			}
+		} else
 			useDefaultDriver();
 
 		File inputFile = cmdr.getFileToSign();
@@ -58,27 +69,52 @@ public class Main {
 		if (!checkFile(inputFile))
 			return;
 
-		 padesSign(inputFile);
+		sign(inputFile);
 
 	}
 
-	private static void padesSign(File inputFile) {
+	private static boolean checkSelectedSignFormat(ArgsParser cmdr) {
+		if (cmdr.isCAdES() && cmdr.isPAdES())
+			return false;
+		else if (cmdr.isCAdES()) {
+			selectedSignFormat = SignFormat.CADES;
+		} else if (cmdr.isPAdES()) {
+			selectedSignFormat = SignFormat.PADES;
+		} else {
+			selectedSignFormat = SignFormat.CADES;
+			System.out.println("Sign Format not provided. Use default format");
+		}
+
+		System.out.println("Selected Sign Format: " + selectedSignFormat.toString());
+		return true;
+	}
+
+	private static void sign(File inputFile) {
 		System.out.println("Start Signature Procedure");
-		char[] pass = getPassword();
+		char[] pass = Utility.getPassword();
 		
-		Pkcs11SignatureToken token = SignService.connectToToken(currentDriverPath, pass);
+		System.out.println();
+
+		AbstractSignFactory factory = null;
+		if (selectedSignFormat == SignFormat.CADES) {
+			factory = new SignFactoryCAdES();
+		} else if (selectedSignFormat == SignFormat.PADES) {
+			factory = new SignFactoryPAdES();
+		}
+
+		Pkcs11SignatureToken token = factory.connectToToken(currentDriverPath, pass);
 		List<DSSPrivateKeyEntry> keys;
 		try {
 			keys = token.getKeys();
 		} catch (DSSException e) {
 			System.err.println("Token access failed.");
-//			e.printStackTrace();
+			// e.printStackTrace();
 			return;
 		}
-		DSSPrivateKeyEntry signer = SignService.getSigner(keys);
+		DSSPrivateKeyEntry signer = factory.getSigner(keys);
 		// Preparing parameters for the PAdES signature
-		PAdESSignatureParameters parameters = SignService.setPAdESParameter(signer);
-		PAdESService service = SignService.createPAdESService();
+		AbstractSignatureParameters parameters = factory.setParameter(signer);
+		AbstractSignatureService service = factory.createService();
 		DSSDocument toSignDocument = new FileDocument(inputFile);
 		// Get the SignedInfo segment that need to be signed.
 		ToBeSigned dataToSign = service.getDataToSign(toSignDocument, parameters);
@@ -91,28 +127,9 @@ public class Main {
 		// obtained the previous step.
 		DSSDocument signedDocument = service.signDocument(toSignDocument, parameters, signatureValue);
 		System.out.println("Creation of signed file");
-		SignService.createSignedPDF(signedDocument, inputFile);
+		factory.createSignedFile(signedDocument, inputFile);
 		System.out.println("END");
 
-	}
-
-	private static char[] getPassword() {
-		Console cnsl = null;
-
-		try {
-			// creates a console object
-			cnsl = System.console();
-			// if console is not null
-			if (cnsl != null) {
-				char[] pwd = cnsl.readPassword("Password: ");
-				return pwd;
-			}
-		} catch (Exception ex) {
-
-			// if any error occurs
-			ex.printStackTrace();
-		}
-		return null;
 	}
 
 	private static boolean checkFile(File inputFile) {
@@ -120,29 +137,32 @@ public class Main {
 			System.err.println("File not exist.");
 			return false;
 		}
-		if (!Files.getFileExtension(inputFile.getName()).equals("pdf")) {
+		if (selectedSignFormat == SignFormat.PADES && !Files.getFileExtension(inputFile.getName()).equals("pdf")) {
 			System.err.println("File is not a pdf.");
 			return false;
 		}
 		return true;
 	}
 
-	private static void setDriver(File file) {
+	private static boolean setDriver(File file) {
 		if (file.exists()) {
 			currentDriverPath = file;
 			System.out.println("Use driver located in: " + currentDriverPath);
+			return true;
 		} else {
-			useDefaultDriver();
+			return useDefaultDriver();
 		}
 	}
 
-	private static void useDefaultDriver() {
-		extractDrivers();
+	private static boolean useDefaultDriver() {
+		SystemType s = null;
 		String os = System.getProperty("os.name").toLowerCase();
-		String arch = checkOSArchitecture();
+		String arch = Utility.checkOSArchitecture();
 		if (os.contains("win")) {
 			currentDriverPath = driverPathWin;
+			s = SystemType.WINDOWS;
 		} else if (os.contains("nix") || os.contains("nux") || os.contains("aix")) {
+			s = SystemType.LINUX;
 			if (arch.equals("64"))
 				currentDriverPath = driverPathLinux64;
 			else
@@ -150,66 +170,53 @@ public class Main {
 		} else if (os.contains("mac")) {
 			// TODO ??
 		}
+		//extract only needed driver
+		if (!extractDrivers(s))
+			return false;
 		System.out.println("Use the default driver located in: " + currentDriverPath);
+		return true;
 
 	}
 
-	private static String checkOSArchitecture() {
-		String arch = System.getenv("PROCESSOR_ARCHITECTURE");
-		String wow64Arch = System.getenv("PROCESSOR_ARCHITEW6432");
-		String realArch = null;
-		if (arch != null && arch.endsWith("64") || wow64Arch != null && wow64Arch.endsWith("64"))
-			realArch = "64";
-		else
-			realArch = "32";
-		return realArch;
-	}
-
-	private static void extractDrivers() {
+	private static boolean extractDrivers(SystemType systype) {
 		try {
-			if (!driverPathWin.exists()) {
+			if (systype == SystemType.WINDOWS && !driverPathWin.exists()) {
 				// load resources
 				URL win = Resources.getResource("resources/driver/Windows/bit4xpki.dll");
-				
+
 				// create folder
-				File wf = new File(buildFilePath("driver", "Windows"));
+				File wf = new File(Utility.buildFilePath("driver", "Windows"));
 				if (!wf.exists())
 					wf.mkdirs();
-				
+
 				// extract resources
 				byte[] bytes = Resources.toByteArray(win);
 				Files.write(bytes, driverPathWin);
 			}
-			if (!driverPathLinux32.exists()) {
+			if (systype == SystemType.LINUX && !driverPathLinux32.exists()) {
 				URL linux32 = Resources.getResource("resources/driver/Linux/32/libbit4xpki.so");
-				File lf = new File(buildFilePath("driver", "Linux", "32"));
+				File lf = new File(Utility.buildFilePath("driver", "Linux", "32"));
 				if (!lf.exists())
 					lf.mkdirs();
 				byte[] bytes = Resources.toByteArray(linux32);
 				Files.write(bytes, driverPathLinux32);
 			}
-			if (!driverPathLinux64.exists()) {
+			if (systype == SystemType.LINUX && !driverPathLinux64.exists()) {
 				URL linux64 = Resources.getResource("resources/driver/Linux/64/libbit4xpki.so");
-				File lf64 = new File(buildFilePath("driver", "Linux", "64"));
+				File lf64 = new File(Utility.buildFilePath("driver", "Linux", "64"));
 				if (!lf64.exists())
 					lf64.mkdirs();
 				byte[] bytes = Resources.toByteArray(linux64);
 				Files.write(bytes, driverPathLinux64);
 			}
+			
+			//TODO add MAC driver
 		} catch (IOException e) {
 			System.err.println("Error in default driver extractaction");
-			e.printStackTrace();
+			return false;
+			// e.printStackTrace();
 		}
-	}
-
-	private static String buildFilePath(String... strings) {
-		String path = "";
-		for (int i = 0; i < strings.length; i++) {
-			path = path + strings[i];
-			if (i != strings.length - 1)
-				path = path + separator;
-		}
-		return path;
+		return true;
 	}
 
 }
